tidying up a bit before a refactor
