+++ .storybook/preview.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Text } from "@jobber/components/Text";
 import { Content } from "@jobber/components/Content";
-import { tokens } from "@jobber/design";
+import { tokens } from "@jobber/design/foundation";
 import { Header } from "./components/Header";
 import { InlineCode } from "./components/InlineCode";
 import { DocsWithSidebar } from "./components/DocsWithSidebar";
@@ -9,8 +9,8 @@ import { CustomCanvas } from "./components/CustomCanvas";
 import { SBProvider } from "./components/SBProvider";
 import { StoryDetails } from "./components/StoryDetails";
 
-import "@atlantis/packages/design/foundation.css";
-import "@atlantis/packages/design/dist/dark.mode.css";
+import "@jobber/design/foundation.css";
+import "@jobber/design/dark.mode.css";
 import "./assets/css/preview.css";
 import { Unstyled } from "@storybook/addon-docs";
 import { Table } from "@storybook/components";
@@ -65,14 +65,14 @@ export const parameters = {
 export const decorators = [
   Story => (
     <div style={{ backgroundColor: 'var(--color-surface)' }}>
-    <SBProvider>
-      <Story />
-    </SBProvider>
+      <SBProvider>
+        <Story />
+      </SBProvider>
     </div>
   ),
-  (Story, context) => 
+  (Story, context) =>
     <div style={{ backgroundColor: 'var(--color-surface)' }}>
-    <StoryDetails Story={Story} context={context} />
+      <StoryDetails Story={Story} context={context} />
     </div>
   ,
 ];
