+++ docs/design/Colors.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta } from "@storybook/addon-docs";
 import { ColorSwatches as Swatch } from "mdxUtils/ColorSwatches";
 import { ColorSwatches } from "@jobber/docx";
-import colors from "@atlantis/packages/design/colors";
+import colors from "@jobber/design/colors";
 
 <Meta title="Design/Colors" />
