+++ packages/components-native/src/Flex/Flex.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { render } from "@testing-library/react-native";
 import { View, ViewStyle } from "react-native";
 import { ReactTestInstance } from "react-test-renderer";
-import { tokens } from "@jobber/design";
+import { tokens } from "@jobber/design/foundation";
 import { Flex } from "./Flex";
 import { FlexProps, Spacing } from "./types";
 import { columnStyles } from "./Flex.styles";
